re-organize __init__.py, separate extensions and remove circular imports from app
