+++ flask/app/__init__.py
@@ -1,13 +1,6 @@
 import logging
 from flask import Flask, logging as flask_logging
-from flask_login import LoginManager
-from flask_migrate import Migrate
-from flask_sqlalchemy import SQLAlchemy
-
-db = SQLAlchemy()
-login = LoginManager()
-migrate = Migrate()
-
+from .extensions import db, login, migrate
 
 from app import (
     buckets,
@@ -19,6 +12,7 @@ from app import (
     analyses,
     groups,
     users,
+    manage,
 )
 
 
@@ -29,28 +23,14 @@ def create_app(config):
 
     # Create the application object
     app = Flask(__name__)
-
     app.config.from_object(config)
 
-    if config.SQLALCHEMY_LOG:
-        logging.basicConfig()
-        logging.getLogger("sqlalchemy.engine").setLevel(logging.INFO)
-        logging.getLogger("sqlalchemy.pool").setLevel(logging.INFO)
-
-    flask_logging.create_logger(app)
-    login.session_protection = "strong"
-
-    db.init_app(app)
-    login.init_app(app)
-    migrate.init_app(app, db)
-
+    register_extensions(app)
+    register_commands(app)
     register_blueprints(app)
+    config_logger(app)
 
-    with app.app_context():
-        # Import here so that other modules have access to the app context
-        from . import manage
-
-        return app
+    return app
 
 
 def register_blueprints(app):
@@ -66,3 +46,24 @@ def register_blueprints(app):
     app.register_blueprint(buckets.bucket_blueprint)
     app.register_blueprint(groups.groups_blueprint)
     app.register_blueprint(users.users_blueprint)
+
+
+def register_extensions(app):
+    db.init_app(app)
+    migrate.init_app(app, db)
+    login.init_app(app)
+
+
+def register_commands(app):
+    app.cli.add_command(manage.add_dummy_data)
+    app.cli.add_command(manage.add_default_admin)
+
+
+def config_logger(app):
+
+    if app.config["SQLALCHEMY_LOG"]:
+        logging.basicConfig()
+        logging.getLogger("sqlalchemy.engine").setLevel(logging.INFO)
+        logging.getLogger("sqlalchemy.pool").setLevel(logging.INFO)
+
+    flask_logging.create_logger(app)
