+++ flask/app/analyses.py
@@ -3,7 +3,8 @@ import json
 from typing import Any, Callable, Dict, List, Union
 from dataclasses import asdict
 
-from . import db, login, models, routes
+from .extensions import db, login
+from . import models
 
 from flask import abort, jsonify, request, Response, Blueprint, current_app as app
 from flask_login import login_user, logout_user, current_user, login_required
@@ -11,6 +12,8 @@ from sqlalchemy import exc
 from sqlalchemy.orm import contains_eager, aliased, joinedload
 from werkzeug.exceptions import HTTPException
 
+
+from .utils import mixin, check_admin, transaction_or_abort
 
 analyses_blueprint = Blueprint(
     "analyses",
@@ -188,7 +191,7 @@ def create_analysis():
     )
 
     db.session.add(obj)
-    routes.transaction_or_abort(db.session.flush)
+    transaction_or_abort(db.session.flush)
 
     # update the dataset_analyses table
     dataset_analyses_obj = [
@@ -202,7 +205,7 @@ def create_analysis():
         db.session.rollback()
         return "Server error", 500
 
-    routes.transaction_or_abort(db.session.commit)
+    transaction_or_abort(db.session.commit)
 
     return jsonify(
         {
@@ -216,7 +219,7 @@ def create_analysis():
 
 @analyses_blueprint.route("/api/analyses/<int:id>", methods=["DELETE"])
 @login_required
-@routes.check_admin
+@check_admin
 def delete_analysis(id: int):
     analysis = models.Analysis.query.filter(
         models.Analysis.analysis_id == id
@@ -291,7 +294,7 @@ def update_analysis(id: int):
             else:
                 return "assignee not found", 400
 
-    enum_error = utils.mixin(analysis, request.json, editable_columns)
+    enum_error = mixin(analysis, request.json, editable_columns)
 
     if enum_error:
         return enum_error, 400
@@ -299,7 +302,7 @@ def update_analysis(id: int):
     if user_id:
         analysis.updated_by_id = user_id
 
-    routes.transaction_or_abort(db.session.commit)
+    transaction_or_abort(db.session.commit)
 
     return jsonify(
         {
