+++ flask/app/families.py
@@ -2,7 +2,8 @@ from dataclasses import asdict
 
 from flask import abort, jsonify, request, Response, Blueprint, current_app as app
 from flask_login import login_user, logout_user, current_user, login_required
-from app import db, login, models
+from .extensions import db, login
+from . import models
 from sqlalchemy.orm import contains_eager, joinedload
 from .utils import check_admin, transaction_or_abort
