+++ flask/app/groups.py
@@ -1,6 +1,7 @@
 from flask import abort, jsonify, request, Response, Blueprint, current_app as app
 from flask_login import login_user, logout_user, current_user, login_required
-from app import db, login, models
+from .extensions import db, login
+from . import models
 from sqlalchemy.orm import contains_eager, joinedload
 from .utils import check_admin, transaction_or_abort, mixin
